Added distortion, array of frequencies in C minor scale, array for the oscillator types, and envelope
The sketch now generates a random oscillator type (sine, square, sawtooth, or triangle) for the carrier signal.

This is pitched at a random note within the C minor scale.

Distortion has been added.

An ADSR envelope has been added that is not yet working.

the carrier amplitudes have been lowered to a max of 0.2 (in the toggle audio section) so as to prevent peaking.